PR asking completed, PR viewing by organization and Adiamat Completed, PR can be viewed with total price using invoice code,  PR pay by Organization NOT completed yet. Trip Started

+++ database/migrations/2024_07_19_084517_create_contract_details_table.php
@@ -23,8 +23,12 @@ return new class extends Migration
             $table->boolean('with_fuel')->default(0);
             $table->boolean('periodic')->default(0);
 
-            $table->decimal('price_contract', 10, 2); // the contract winning price                                                                            // the driver or vehicle_supplier should not see this price
-            $table->decimal('price_vehicle_payment', 10, 2); // the price that is to be paid for vehicle_supplier or driver (i.e payed to the vehicle account) // the organization should not see this price
+            // 
+            $table->integer('price_contract'); // the contract winning price                                                                            // the driver or vehicle_supplier should not see this price
+            $table->integer('price_vehicle_payment'); // the price that is to be paid for vehicle_supplier or driver (i.e payed to the vehicle account) // the organization should not see this price
+            // the above integer is more suitable since we are ignoring cents
+            // $table->decimal('price_contract', 10, 2); 
+            // $table->decimal('price_vehicle_payment', 10, 2); 
 
             $table->decimal('tax', 4, 2)->default(ContractDetail::CONTRACT_DETAIL_DEFAULT_TAX_15);
